rename homeComponents to component
