+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/screen/home/component/Title.kt
@@ -1,4 +1,4 @@
-package com.sanaa.tudee_assistant.presentation.screen.home.homeComponents
+package com.sanaa.tudee_assistant.presentation.screen.home.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
